Splitting apart DefaultItemViewWithProvenance
